Extract the sidebar Queries panel into SavedQueriesPanel UserControl
Second step of the MainWindow decomposition (UI design rule 7). Moves the
saved-queries + history sidebar tab's markup, its double-tap-to-open wiring,
and the OnQueryListDoubleTapped helper out of MainWindow into a focused
UserControl bound to the existing SavedQueriesViewModel via DataContext. The
double-tap now resolves the VM from DataContext instead of the host's
_viewModel field. No behavior change. Drops the now-unused query: xmlns from
MainWindow.axaml.
